TF: First code commit
This is the first round of code for this module.

The aim is to create an AWS ACM Certificate and validate it either
by DNS (and create the DNS records in Route 53) or by email and then
return the ARN of the certificate so that it can be used by either
CloudFront or by a Load Balancer.

Use single list of maps for inputing all domains and zone ids

Fix var names

Fix description

Fix lookup

Fix example

Updating of descriptions, documentation and tests

Make documentation a little neater

terraform fmt code in README

Refinements

Allowing the output to be empty

Fix test output

Simplify the cert validation process

Refactor passing of the AWS Region and outputs

Updating the README

resource refactor for neatness

Rename certificate arn output

Fix variable descriptions, update README and remove unused variable

Refine output description and README

Remove non-required stars